Corrected a spelling mistake in Controller.cpp and an incorrect comment in ProMini.ino Also fixed the type of a variable

+++ ProMini/ProMini.h
@@ -136,7 +136,7 @@ class SpeedSensor : public Sensor {
 class Controller
 {
   private:
-    int updateInterval;      // interval between updates
+    unsigned long updateInterval;      // interval between updates
     unsigned long lastUpdate; // last update of position
     Transducer &transducer;
     Sensor &sensor;
